// src/gql/types/typedSchema.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type Ad = {
  __typename?: 'Ad';
  description?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  imageList?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  location?: Maybe<Scalars['String']['output']>;
  phoneNumber?: Maybe<Scalars['String']['output']>;
  promoteImage?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type Mutation = {
  __typename?: 'Mutation';
  createAd?: Maybe<Ad>;
  deletePartner?: Maybe<Scalars['Boolean']['output']>;
  registerPartner?: Maybe<Partner>;
};


export type MutationCreateAdArgs = {
  input?: InputMaybe<AdInput>;
};


export type MutationDeletePartnerArgs = {
  input?: InputMaybe<DeletePartnerInput>;
};


export type MutationRegisterPartnerArgs = {
  input?: InputMaybe<PartnerInput>;
};

export type Partner = {
  __typename?: 'Partner';
  email: Scalars['String']['output'];
  id?: Maybe<Scalars['ID']['output']>;
  name: Scalars['String']['output'];
  username?: Maybe<Scalars['String']['output']>;
};

export type PartnerInput = {
  email: Scalars['String']['input'];
  name: Scalars['String']['input'];
  password: Scalars['String']['input'];
  username?: InputMaybe<Scalars['String']['input']>;
};

export type Query = {
  __typename?: 'Query';
  getAd: Ad;
  getAds?: Maybe<Array<Maybe<Ad>>>;
  login?: Maybe<Partner>;
};


export type QueryGetAdArgs = {
  id: Scalars['ID']['input'];
};


export type QueryLoginArgs = {
  input: LoginInput;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;



/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Ad: ResolverTypeWrapper<Ad>;
  AdInput: AdInput;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  DeletePartnerInput: DeletePartnerInput;
  ID: ResolverTypeWrapper<Scalars['ID']['output']>;
  LoginInput: LoginInput;
  Mutation: ResolverTypeWrapper<{}>;
  Partner: ResolverTypeWrapper<Partner>;
  PartnerInput: PartnerInput;
  Query: ResolverTypeWrapper<{}>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Ad: Ad;
  AdInput: AdInput;
  Boolean: Scalars['Boolean']['output'];
  DeletePartnerInput: DeletePartnerInput;
  ID: Scalars['ID']['output'];
  LoginInput: LoginInput;
  Mutation: {};
  Partner: Partner;
  PartnerInput: PartnerInput;
  Query: {};
  String: Scalars['String']['output'];
};

export type AdResolvers<ContextType = any, ParentType extends ResolversParentTypes['Ad'] = ResolversParentTypes['Ad']> = {
  description?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<Maybe<ResolversTypes['ID']>, ParentType, ContextType>;
  imageList?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
  location?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  phoneNumber?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  promoteImage?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  title?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  createAd?: Resolver<Maybe<ResolversTypes['Ad']>, ParentType, ContextType, Partial<MutationCreateAdArgs>>;
  deletePartner?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType, Partial<MutationDeletePartnerArgs>>;
  registerPartner?: Resolver<Maybe<ResolversTypes['Partner']>, ParentType, ContextType, Partial<MutationRegisterPartnerArgs>>;
};

export type PartnerResolvers<ContextType = any, ParentType extends ResolversParentTypes['Partner'] = ResolversParentTypes['Partner']> = {
  email?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  id?: Resolver<Maybe<ResolversTypes['ID']>, ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  username?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  getAd?: Resolver<ResolversTypes['Ad'], ParentType, ContextType, RequireFields<QueryGetAdArgs, 'id'>>;
  getAds?: Resolver<Maybe<Array<Maybe<ResolversTypes['Ad']>>>, ParentType, ContextType>;
  login?: Resolver<Maybe<ResolversTypes['Partner']>, ParentType, ContextType, RequireFields<QueryLoginArgs, 'input'>>;
};

export type Resolvers<ContextType = any> = {
  Ad?: AdResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
  Partner?: PartnerResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
};